% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoRch.generator-package.R
\docType{package}
\name{scoRch.generator-package}
\alias{scoRch.generator}
\alias{scoRch.generator-package}
\title{scoRch.generator: Spatiotemporal Classification of Regional Compound Heatwaves}
\description{
Detects local heatwaves in gridded daily maximum temperature, selects regionally extensive heatwave days, groups their heatwave cells with DBSCAN, fits principal-component ellipses with temperature-weighted centroids, classifies consecutive-day compound heatwave events into four space-time types, and fits the power-law and log-Gaussian Cox process models of the SCORCH method of Bouhamad and Najibi. One configuration file and one function run the complete workflow and write the catalogs, tables and descriptively named result plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fawazbouhamad/scoRch.generator}
  \item Report bugs at \url{https://github.com/fawazbouhamad/scoRch.generator/issues}
}

}
\author{
\strong{Maintainer}: Fawaz Bouhamad \email{fawawaz@hotmail.com}

Authors:
\itemize{
  \item Fawaz Bouhamad \email{fawawaz@hotmail.com}
  \item Nasser Najibi
}

}
\keyword{internal}
